Pets

+++ src/main/java/com/leverx/service/impl/CatServiceImpl.java
@@ -3,15 +3,13 @@ package com.leverx.service.impl;
 import static com.leverx.dto.converter.CatConverterDto.*;
 import static com.leverx.entity.EPetType.CAT;
 
-
 import java.util.List;
-import java.util.Optional;
+
+import javax.persistence.EntityNotFoundException;
 
 import com.leverx.dto.request.CatRequestDto;
 import com.leverx.dto.response.CatResponseDto;
 import com.leverx.entity.Cat;
-import com.leverx.exception.CatNotFoundException;
-import com.leverx.exception.UserNotFoundException;
 import com.leverx.repository.CatRepository;
 import com.leverx.repository.UserRepository;
 import com.leverx.service.CatService;
@@ -41,7 +39,7 @@ public class CatServiceImpl implements CatService {
     cat.setOwner(
         userRepository
             .findById(catRequestDto.getOwnerId())
-            .orElseThrow(UserNotFoundException::new));
+            .orElseThrow(() -> new EntityNotFoundException("There is no such user")));
     cat.setPetType(CAT);
 
     return convertCatEntityToResponse(catRepository.save(cat));
@@ -55,18 +53,24 @@ public class CatServiceImpl implements CatService {
     cat.setOwner(
         userRepository
             .findById(catRequestDto.getOwnerId())
-            .orElseThrow(UserNotFoundException::new));
-    cat.setId(catRepository.findById(catId).orElseThrow(CatNotFoundException::new).getId());
+            .orElseThrow(() -> new EntityNotFoundException("There is no such user")));
+    cat.setId(
+        catRepository
+            .findById(catId)
+            .orElseThrow(() -> new EntityNotFoundException("There is no such cat"))
+            .getId());
+    cat.setPetType(CAT);
 
     return convertCatEntityToResponse(catRepository.save(cat));
   }
 
   @Override
   @Transactional
-  public Optional<CatResponseDto> findById(final long id) {
-    return Optional.of(convertCatEntityToResponse(
-        catRepository.findById(id)
-                .orElseThrow(CatNotFoundException::new)));
+  public CatResponseDto findById(final long id) {
+    return convertCatEntityToResponse(
+        catRepository
+            .findById(id)
+            .orElseThrow(() -> new EntityNotFoundException("There is no such cat")));
   }
 
   @Override
@@ -78,6 +82,9 @@ public class CatServiceImpl implements CatService {
   @Override
   @Transactional
   public void delete(final long id) {
+    if (!catRepository.existsById(id)) {
+      throw new EntityNotFoundException("There is no such dog");
+    }
     catRepository.deleteById(id);
   }
 }
